requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
PyPDF2>=3.0.0
python-docx>=0.8.0
openpyxl>=3.0.0
requests>=2.28.0
python-dotenv>=1.0.0
